Abs: wMenuCreate splitted: Create and map
The wMenuCreate function is splitted in two functions, one to create
the menu, without the screen, and other to map the menu on the screen.

The menu core and the menu frame is created together, at the create
function.

Because the frame core width is 1, we can replace it in the wcore
creation. Now, we use the "width" variable to do it.