Create one Dock
This patch changes the current behaviour, one dock per screen, with
only one dock for all screens.

This patch is only as a transition patch, because we can handle
multiple docks.

The problem is the Dock is attached to the WScreen, so the Dock is
not screen independent. This patch moves the Dock to the global
preferences, and then the dock could be created without screens.
If the screen changes, the dock doesn't need be changed.

Future patches should allow create more than one dock, using a
dock array in the global preferences variable. We can have multiple
docks if we have multiple screens, if the user want's it.