FW: Code for udpateTexture_resizebar
This patch moves the udpate texture code from remakeTexture_resizebar to
udpateTexture_resizebar.

In the previous code, when the function remakeTexture_resizebar is used,
then udpateTexture_resizebar is called, except in the for loop, that we
need to include the "if state is 0".

Now the code has more logic, remakeTexture remakes the texture and
updateTexture updates it.