FWA: Check resizebar flag
This patch includes the "Check resizebar flag" when the resizebar is
used. Then, when the flag is true, the resizebar is mapped, and when
the resizebard is unmapped, the flag is flase.

This patch will used in the next patch when the resizebar exits always.
This patch is the first patch in the FWA (FrameWin Abstraction) set.