FWA: Resizebar abstraction
This patch includes the resizebar in the framewin code.

Before this patch, the resizebar is created and destroyed when the
user wants, not only hidden. This behaviour includes some troubles.

For example, launch a XTerm or other application and then open the
Attributes window (right click on titlebar). Then select "Disable
resizebar" and click in Apply. The resizebar is destroyed. Then select
"Disable resizebar" again and click Apply. The resizebar is created
again. Then, select "Disable resizebar" and click Apply. Windowmaker
crash.

With this patch the resizebar is created, mapped, unmapped, destroyed,
but is always created when the framewin is created and destroyed when
the framewin is destroyed. The user only can map or unmap it.

This patch includes the functions to do these actions (map, unmap,...)
and the code was moved from the original functions. No new lines included
and no lines removed.