Remove menu brother
Menu brother is used when we launch the menu, then select the menu
bar. The close button (X) is painted. When we launch the menu again
we are playing with the original menu, and the brother.

This is a wrong behaviour, because we are playing with two different
menus, but we need only one and paint it twice or MORE times.

We need destroy first the brother menu and play with the map/unmap
in next patches.

This patch also changes some function prototypes, because the arguments
are not longer used. I did it here to avoid compiler warnings.

This patch is based in remove the calls, functions,... where the
brother is used. I did it looking for "brother" and removing the code.
I didn't have problems and is easy to understand, but is a big patch.

The only problem was with the function closeBrotherCascadesOf(). This
function closes automatically the menu when the right mouse is released,
so the behaviour was different. I found the problem and I changed this
lines:

@@ -1908,16 +1892,13 @@ static void menuMouseDown(WObjDescriptor *desc, XEvent *event)
                        selectEntry(menu, -1);
                } else {
                        if (entry->cascade >= 0 && menu->cascades)
-                               selectEntry(menu, -1);
+                               selectEntry(menu, entry_no);
                }
        }

        if (close_on_exit || !smenu)
                closeCascade(desc->parent);

-       /* close the cascade windows that should not remain opened */
-       closeBrotherCascadesOf(desc->parent);
-

And the problems is gone. This is the only "rare" change.