paint_title don't use all_buttons argument
The function paint_title is using the "all_buttons" argument to calculate
the right position for the title. This variable is not needed if we use
the right and left space variables.

Now, the code is a little bit more slow, because when all buttons are
painted, the paint_title function needs calculate the position in a
little bit more complex function (no x/2). But the code is cleaner now.