D: Clip painting use WDock coords
The Clip painting was using the AppIcon coords to paint the Clip,
but we can use the WDock session saved directly to paint the Clip.

This patch init the WDock coords to 0 when the Clip is created.
Then when the saved session coords are loaded, we update the
AppIcon coords with this value.

Now, restoreClip don't need set up the Clip coords.