DW: Use restore_state_autocollapsed in drawer
The function restore_state_autocollapsed could be used for dock and for
drawers, so we can change the code.

We need change the restore_state_autocollapsedto return a value. This
value is used to change the default value to 0. Probably we can change
it in the function creation, but I am not sure yet.