IA: Remove wAppIconCreateForDock
This patch is "rare". It removes the function wAppIconCreateForDock and
moves their contents four times! to the places where the function was used.

Yes, it seems to be bad, but it has their logic. I will remove some lines
in the next patches and I will create abstraction between creation and mapping.

Please, be patient ;-)