D: Remove early clip creation
This patch removes the main icon creation in the screen.h file. Now,
the icon is created when the Clip is created (when the workspace is
created).

About the Session state options (DND and Paste), the session options
are loaded just before the clip creation, so we can remove them here.
If we have problems with the session options, then we can add them
to these functions, after the clip creation.