D: Merge AppIcon and Icon creation
This patch merges the icon creation and the appicon creation in the
same function. Now, the specific funtions to create clip, dock and
drawer creates the clip, dock and drawer struct and their specific
icon.

The aim of this patch is hold all together, and then set the possion
info only one time. I will do this work in next patches.