WM: Move the last_dock variable to w_global
This patch moves the "last_dock" variable to w_global.

Moving it, some WScreen variables are not longer used and some
function prototypes change for it.

This patch removes the last_dock variable per WScreen, and
uses the same last_dock for all screens now. This behavior
could be changed in the future, but I am not sure what we need.