WM: Move drawer info to w_global
This patch moves the Drawer Chain, the drawer_count and the
selected attracting drawer to the struct w_global in WindowMaker.h.

This patch changes a lot of files, because these variables are used
in all of them. The patch also change some function prototypes, because
the WScreen variable sometimes is not used now.

Finally, I did some style changes, because I won't create an extra
patch only for these style issues.