FWA: Unmap is not destroy 1 of 3
This patch is the first of three patches.

The aim of this set is split the creation plus mapping and the
unmapping plus destroy calls.

This first patch only adds the check for the titlebar and resizebar
flags, because in the next patches in this serie, the titlebar and
resizebar structs will be exist always, but sometimes will be painted
and sometimes not.

We will use the flags.titlebar and flags.resizebar to check if the
titlebar or resizebar are mapped (painted).