New function dock_icon_create_core
The function dock_icon_create_core includes the icon creation. Moves
the code from dock_icon_create.

Now, the function dock_icon_create only creates the icon and then
add it to the icon list.

This patch is created because we can use this function in the function
wAppIcon_create function, that do the same, but without incluce the icon
in the icon list. A in next patches.