DA: Split drawerRestoreState set and mapping
This patch split the function drawerRestoreState in the functions
drawerRestoreState and drawerRestoreState_map.

The function drawerRestoreState only set the values, the function
drawerRestoreState_map map the dock and their docks apps in the screen.

We can remove the apps "if" check in the function drawerRestoreState_map
because the map_set_attacheddocks function checks if the icon is NULL.