WSA: Make the Clip creation similar
This patch makes the Clip creation similar in the functions
wWorkspaceNew and wWorkspaceRestoreState.

The Clip creation in the function wWorkspaceNew checks if the workspace
number is 0, so is the first workspace, and call only for that workspace
the clip icon creation. In the icon creation, we check if the icon image
was previously created.

Really, we can avoid this check, and call the icon creation always. If the
icon was previosly created, then the function won't do nothing.

This patch also moves the get from dictionary call some lines above.