VSCR: Move clip info to virtual_screen
This patch moves all the related info about clip to the virtual_screen.
Now, every virtual screen will have their own clip, and the clip
is not shared between phisical screens (in Xinerama there is only one clip).

The Clip icon is shared, so is included in the global info.