VSCR: Virtual Screen is now pointer
This patch creates the virtual_screen as pointer in the Screen
Initialization (wScreenInit). When the screen is created, the
virtual screen is created too.

This patch includes a new pointer in virtual_screen to the screen,
for this reason I need move the typedefs for virtual_screen and
wScreen in Windowmaker.h. The pointer is set in wScreenInit.

The changes in the other files is only for struct to struct pointer
movement.